Refactor controller column
- move callsign txt to assets
